remove a lot of functions from public interface of utils
